feat : controll lotto game

+++ src/main/java/ui/PrintService.java
@@ -5,5 +5,5 @@ import java.util.Map;
 
 public interface PrintService {
     void printLotto(List<List<Integer>> lottoList);
-    void printWinningResult(Map<Rank, Integer> result, long yield);
+    void printWinningResult(Map<Rank, Integer> result, double yield);
 }
